+++ config/application.rb
@@ -26,7 +26,7 @@ module DC
     # config.i18n.default_locale = :de
 
     # DocumentCloud-specific configuration.
-    ::DC::SECRETS.merge!   YAML.load_file("#{Rails.root}/secrets/secrets.yml")[Rails.env]
+    ::DC::SECRETS.merge!   YAML.load_file(Rails.root.join('config','server', 'secrets/secrets.yml'))[Rails.env]
     ::DC::CONFIG.merge! YAML.load( ERB.new(File.read( Rails.root.join('config','document_cloud.yml') ) ).result )[Rails.env]
 
     ::DC::ANALYTICS_DB = YAML.load(ERB.new(File.read("#{Rails.root}/config/database_analytics.yml")).result(binding))[Rails.env]
