+++ config/database.yml
@@ -1,4 +1,4 @@
-<% secrets = YAML.load_file(File.join(Rails.root, 'secrets/secrets.yml'))[ENV['RAILS_ENV'] || 'development'] %>
+<% secrets = DC::SECRETS %>
 
 defaults: &defaults
   adapter: postgresql
