+++ config/database_analytics.yml
@@ -1,4 +1,4 @@
-<% secrets = YAML.load_file('secrets/secrets.yml')[ENV['RAILS_ENV'] || 'development'] %>
+<% secrets = DC::SECRETS %>
 
 defaults: &defaults
   adapter: postgresql
