+++ config/document_cloud.yml
@@ -1,4 +1,4 @@
-<% secrets = YAML.load_file('./secrets/secrets.yml')[ENV['RAILS_ENV'] || "development"] %>
+<% secrets = DC::SECRETS %>
 defaults: &defaults
   cloud_crowd_server:   http://dev.dcloud.org:8080
   server_root:          dev.dcloud.org
